requirements.txt
ply==3.11
python-dotenv==1.0.0
openai==0.28.0
graphviz==0.20.1